# perf: reduce branches