As Bob Marley said; every little thing is gonna be alright.

## Before

```
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark           Time             CPU   Iterations        GHz cycle/byte cycles/mime instructions/byte instructions/cycle instructions/mime instructions/ns     mime/s    ns/mime      speed  time/byte  time/mime
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench      72347 ns        72346 ns         9677    3.27756    23.7897     268.168           128.236            5.39037          1.44553k         17.6672 12.1775M/s    81.8195  137.27M/s  7.28492ns  82.1186ns

```

## After

```
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark           Time             CPU   Iterations        GHz cycle/byte cycles/mime instructions/byte instructions/cycle instructions/mime instructions/ns     mime/s    ns/mime      speed  time/byte  time/mime
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench      68337 ns        68337 ns        10265    3.27956    22.4421     252.976           119.032            5.30399          1.34178k         17.3947  12.892M/s    77.1373 145.324M/s   6.8812ns  77.5677ns

```