@anonrig 2023-06-27
@KhafraDev @lemire The last commit I've pushed for adding collecting code points, was missing from our code, which was adding unnecessary overhead (and re-iteration)